alot changes

+++ app/src/main/java/com/example/allegroandroid/ia/posedetector/PoseEvaluatedResult.java
@@ -1,11 +1,11 @@
 package com.example.allegroandroid.ia.posedetector;
 
 public class PoseEvaluatedResult {
-   public Integer confidence;
-   public  String resultado;
+   public Double confidence;
+   public String resultado;
    public Integer colorResult, colorBackGround;
 
-    public PoseEvaluatedResult(int confidence, String resultado,  int colorResult, int colorBackGround) {
+    public PoseEvaluatedResult(double confidence, String resultado,  int colorResult, int colorBackGround) {
         this.confidence = confidence * 100;
         this.resultado = resultado;
         this.colorBackGround = colorBackGround;
